requirements.txt
fastapi
uvicorn
python-multipart
easyocr
opencv-python-headless
numpy
python-dotenv
aiofiles